Moved RPC to new internal folder, init protobuf support.
