+++ cli/terminal.go
@@ -14,13 +14,13 @@ import (
 
 	"github.com/fatih/color"
 	"github.com/mitsukomegumi/GoP2P/common"
-	commonProto "github.com/mitsukomegumi/GoP2P/rpc/proto/common"
-	databaseProto "github.com/mitsukomegumi/GoP2P/rpc/proto/database"
-	environmentProto "github.com/mitsukomegumi/GoP2P/rpc/proto/environment"
-	handlerProto "github.com/mitsukomegumi/GoP2P/rpc/proto/handler"
-	nodeProto "github.com/mitsukomegumi/GoP2P/rpc/proto/node"
-	shardProto "github.com/mitsukomegumi/GoP2P/rpc/proto/shard"
-	upnpProto "github.com/mitsukomegumi/GoP2P/rpc/proto/upnp"
+	commonProto "github.com/mitsukomegumi/GoP2P/internal/rpc/proto/common"
+	databaseProto "github.com/mitsukomegumi/GoP2P/internal/rpc/proto/database"
+	environmentProto "github.com/mitsukomegumi/GoP2P/internal/rpc/proto/environment"
+	handlerProto "github.com/mitsukomegumi/GoP2P/internal/rpc/proto/handler"
+	nodeProto "github.com/mitsukomegumi/GoP2P/internal/rpc/proto/node"
+	shardProto "github.com/mitsukomegumi/GoP2P/internal/rpc/proto/shard"
+	upnpProto "github.com/mitsukomegumi/GoP2P/internal/rpc/proto/upnp"
 )
 
 // Terminal - absctract container holding set of variable with values (runtime only)
