+++ common/common.go
@@ -32,8 +32,8 @@ const (
 	// NodeAvailableRep - global definition for reputation value of node availability
 	NodeAvailableRep = 10
 
-	// GoP2PTestNetID - GoP2P testing network identifier
-	GoP2PTestNetID = 4519161392015
+	// GoP2PTestnetID - GoP2P testing network identifier
+	GoP2PTestnetID = 4519161392015
 
 	// ConnectionDelimiter - GoP2P standard connection delimiter
 	ConnectionDelimiter = byte('\f')
