+++ internal/rpc/common/server.go
@@ -7,7 +7,7 @@ import (
 	"strconv"
 
 	"github.com/mitsukomegumi/GoP2P/common"
-	commonProto "github.com/mitsukomegumi/GoP2P/rpc/proto/common"
+	commonProto "github.com/mitsukomegumi/GoP2P/internal/rpc/proto/common"
 )
 
 // Server - GoP2P RPC server
