+++ internal/rpc/proto/generateFromProto.sh
