+++ internal/rpc/upnp/server.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"strconv"
 
-	upnpProto "github.com/mitsukomegumi/GoP2P/rpc/proto/upnp"
+	upnpProto "github.com/mitsukomegumi/GoP2P/internal/rpc/proto/upnp"
 	"github.com/mitsukomegumi/GoP2P/upnp"
 )
