+++ main.go
@@ -10,20 +10,20 @@ import (
 	"github.com/fatih/color"
 	"github.com/mitsukomegumi/GoP2P/cli"
 	"github.com/mitsukomegumi/GoP2P/common"
-	commonServer "github.com/mitsukomegumi/GoP2P/rpc/common"
-	"github.com/mitsukomegumi/GoP2P/rpc/database"
-	"github.com/mitsukomegumi/GoP2P/rpc/environment"
-	handlerServer "github.com/mitsukomegumi/GoP2P/rpc/handler"
-	nodeServer "github.com/mitsukomegumi/GoP2P/rpc/node"
-	commonProto "github.com/mitsukomegumi/GoP2P/rpc/proto/common"
-	databaseProto "github.com/mitsukomegumi/GoP2P/rpc/proto/database"
-	environmentProto "github.com/mitsukomegumi/GoP2P/rpc/proto/environment"
-	handlerProto "github.com/mitsukomegumi/GoP2P/rpc/proto/handler"
-	nodeProto "github.com/mitsukomegumi/GoP2P/rpc/proto/node"
-	shardProto "github.com/mitsukomegumi/GoP2P/rpc/proto/shard"
-	upnpProto "github.com/mitsukomegumi/GoP2P/rpc/proto/upnp"
-	shardServer "github.com/mitsukomegumi/GoP2P/rpc/shard"
-	upnpServer "github.com/mitsukomegumi/GoP2P/rpc/upnp"
+	commonServer "github.com/mitsukomegumi/GoP2P/internal/rpc/common"
+	"github.com/mitsukomegumi/GoP2P/internal/rpc/database"
+	"github.com/mitsukomegumi/GoP2P/internal/rpc/environment"
+	handlerServer "github.com/mitsukomegumi/GoP2P/internal/rpc/handler"
+	nodeServer "github.com/mitsukomegumi/GoP2P/internal/rpc/node"
+	commonProto "github.com/mitsukomegumi/GoP2P/internal/rpc/proto/common"
+	databaseProto "github.com/mitsukomegumi/GoP2P/internal/rpc/proto/database"
+	environmentProto "github.com/mitsukomegumi/GoP2P/internal/rpc/proto/environment"
+	handlerProto "github.com/mitsukomegumi/GoP2P/internal/rpc/proto/handler"
+	nodeProto "github.com/mitsukomegumi/GoP2P/internal/rpc/proto/node"
+	shardProto "github.com/mitsukomegumi/GoP2P/internal/rpc/proto/shard"
+	upnpProto "github.com/mitsukomegumi/GoP2P/internal/rpc/proto/upnp"
+	shardServer "github.com/mitsukomegumi/GoP2P/internal/rpc/shard"
+	upnpServer "github.com/mitsukomegumi/GoP2P/internal/rpc/upnp"
 	"github.com/mitsukomegumi/GoP2P/types/handler"
 	"github.com/mitsukomegumi/GoP2P/types/node"
 	"github.com/mitsukomegumi/GoP2P/upnp"
