+++ types/database/database_test.go
@@ -22,7 +22,7 @@ func TestNewDatabase(t *testing.T) {
 	}
 
 	if err == nil {
-		db, err := NewDatabase(node, "GoP2P_TestNet", common.GoP2PTestNetID, 10, "test") // Create new database with bootstrap node, and acceptable timeout
+		db, err := NewDatabase(node, "GoP2P_TestNet", common.GoP2PTestnetID, 10, "test") // Create new database with bootstrap node, and acceptable timeout
 
 		if err != nil && !strings.Contains(err.Error(), "socket") { // Check for errors
 			t.Errorf(err.Error()) // Fail with errors
@@ -44,7 +44,7 @@ func TestAddNode(t *testing.T) {
 		t.FailNow()           // Panic
 	}
 
-	db, err := NewDatabase(node, "GoP2P_TestNet", common.GoP2PTestNetID, 10, "test") // Create new node database with bootstrap node
+	db, err := NewDatabase(node, "GoP2P_TestNet", common.GoP2PTestnetID, 10, "test") // Create new node database with bootstrap node
 
 	if err != nil && !strings.Contains(err.Error(), "socket") { // Check for errors
 		t.Errorf(err.Error()) // Fail with errors
@@ -72,7 +72,7 @@ func TestRemoveNode(t *testing.T) {
 		t.FailNow()           // Panic
 	}
 
-	db, err := NewDatabase(node, "GoP2P_TestNet", common.GoP2PTestNetID, 10, "test") // Create new node database with bootstrap node
+	db, err := NewDatabase(node, "GoP2P_TestNet", common.GoP2PTestnetID, 10, "test") // Create new node database with bootstrap node
 
 	if err != nil && !strings.Contains(err.Error(), "socket") { // Check for errors
 		t.Errorf(err.Error()) // Fail with errors
@@ -100,7 +100,7 @@ func TestQueryForAddress(t *testing.T) {
 		t.FailNow()           // Panic
 	}
 
-	db, err := NewDatabase(node, "GoP2P_TestNet", common.GoP2PTestNetID, 10, "test") // Create new node database with bootstrap node
+	db, err := NewDatabase(node, "GoP2P_TestNet", common.GoP2PTestnetID, 10, "test") // Create new node database with bootstrap node
 
 	if err != nil && !strings.Contains(err.Error(), "socket") { // Check for errors
 		t.Errorf(err.Error()) // Log found error
@@ -128,7 +128,7 @@ func TestUpdateRemoteDatabase(t *testing.T) {
 		t.FailNow()           // Panic
 	}
 
-	db, err := NewDatabase(node, "GoP2P_TestNet", common.GoP2PTestNetID, 10, "test") // Create new node database with bootstrap node
+	db, err := NewDatabase(node, "GoP2P_TestNet", common.GoP2PTestnetID, 10, "test") // Create new node database with bootstrap node
 
 	if err != nil && !strings.Contains(err.Error(), "socket") { // Check for errors
 		t.Errorf(err.Error()) // Log found error
@@ -156,7 +156,7 @@ func TestJoinDatabase(t *testing.T) {
 		t.FailNow()           // Panic
 	}
 
-	db, err := NewDatabase(node, "GoP2P_TestNet", common.GoP2PTestNetID, 10, "test") // Create new node database with bootstrap node
+	db, err := NewDatabase(node, "GoP2P_TestNet", common.GoP2PTestnetID, 10, "test") // Create new node database with bootstrap node
 
 	if err != nil && !strings.Contains(err.Error(), "socket") && !strings.Contains(err.Error(), "timed out") { // Check for errors
 		t.Errorf(err.Error()) // Log found error
@@ -188,7 +188,7 @@ func TestFetchRemoteDatabase(t *testing.T) {
 		t.FailNow()           // Panic
 	}
 
-	_, err = NewDatabase(node, "GoP2P_TestNet", common.GoP2PTestNetID, 10, "test") // Create new node database with bootstrap node
+	_, err = NewDatabase(node, "GoP2P_TestNet", common.GoP2PTestnetID, 10, "test") // Create new node database with bootstrap node
 
 	if err != nil && !strings.Contains(err.Error(), "socket") { // Check for errors
 		t.Errorf(err.Error()) // Log found error
